In this branch TaskBroker (DatacenterBroker) is allowed to communicate with VM scheduler and Cloudlet scheduler via datacenter.
TaskBroker is no longer allowed to communicate with these scheduler by
calling public methods of them.

The communication between them is encapsulated in messages so that all
kinds of requests and and replies can be transmitted in a united way.

This is a stash commit. Further commits are necessary to complete the
functionality.